% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenario_outputs}
\alias{scenario_outputs}
\alias{scenario_target_ts}
\title{Scenario outputs}
\format{
\subsection{\code{scenario_outputs}}{

A data frame with 7,176 rows and 9 columns:
\describe{
\item{model_id}{the name of the model}
\item{origin_date}{the starting point of the projection in yyyy-mm-dd
format}
\item{scenario_id}{a unique identifier for the scenario}
\item{location}{FIPS code identifying a location}
\item{target}{a unique identifier for the target}
\item{horizon}{number of time units ahead being projected relative to the
\code{origin_date}, in units of weeks}
\item{output_type}{the type of representation of the prediction}
\item{output_type_id}{more identifying information specific to the
output type; output_type_id\verb{is not relevant for every kind of}output_type\verb{(for example, hubs will not expect}output_type_id\verb{values when the}output_type` is mean or median}
\item{value}{the model’s prediction}

}
}

\subsection{\code{scenario_target_ts}}{

A data frame with 10,255 rows and 3 columns:
\describe{
\item{date}{the date of the target observation in yyyy-mm-dd format}
\item{location}{FIPS code identifying a location}
\item{value}{a count of incident cases in the given \code{location}
in the week ending on the given \code{date}.}
\item{target}{a unique identifier for the target}
}
}
}
\usage{
scenario_outputs

scenario_target_ts
}
\description{
\code{scenario_outputs} contains example scenario projection data that represents
model outputs and an ensemble from a scenario hub with predictions for one
target (\verb{inc hosp}) in one location (\code{"US"}), one round ("2021-03-07") and
four scenario. This dataset reflect scenario projection outputs as they look
when retrieved from a hub via the \code{hubData} package (which is slightly
different than they look when originally submitted by modelers), and with
a "mean" ensemble calculated with the \code{hubEnsemble} package
\code{simple_ensemble()} function with default parameters.
The date of occurrence of the outcome of interest in yyyy-mm-dd format
can be calculated directly from the \code{origin_date} and \code{horizon} as follows:
\code{target_end_date = origin_date + (7 * horizon) - 1}

\code{scenario_target_ts} contains time series target data associated with the
scenario projection data.
}
\keyword{datasets}
